Code Review

+++ tests/system/feature_store/test_spark_engine.py
@@ -1,4 +1,3 @@
-# Marked as enterprise because of v3io mount and pipelines
 import os
 
 import pytest
@@ -10,17 +9,21 @@ from tests.system.base import TestMLRunSystem
 
 
 @TestMLRunSystem.skip_test_if_env_not_configured
+# Marked as enterprise because of v3io mount and remote spark
 @pytest.mark.enterprise
 class TestFeatureStoreSparkEngine(TestMLRunSystem):
     project_name = "fs-system-spark-engine"
     spark_service = ""
     pq_source = "testdata.parquet"
+    spark_image_deployed = (
+        False  # Set to True if you want to avoid the image building phase
+    )
+    test_branch = ""  # For testing specific branche. e.g.: "https://github.com/mlrun/mlrun.git@development"
 
     @classmethod
     def _init_env_from_file(cls):
         env = cls._get_env_from_file()
-        cls.spark_service = env["DEFAULT_SPARK_SERVICE"]
-        os.environ.setdefault("V3IO_USERNAME", env["V3IO_USERNAME"])
+        cls.spark_service = env["MLRUN_SYSTEM_TESTS_DEFAULT_SPARK_SERVICE"]
 
     def get_local_pq_source_path(self):
         return os.path.relpath(str(self.assets_path / self.pq_source))
@@ -37,7 +40,7 @@ class TestFeatureStoreSparkEngine(TestMLRunSystem):
         from mlrun.run import new_function
         from mlrun.runtimes import RemoteSparkRuntime
 
-        if not self.spark_service:
+        if not self.spark_image_deployed:
             self._init_env_from_file()
 
             store, _ = store_manager.get_or_create_store(
@@ -47,18 +50,19 @@ class TestFeatureStoreSparkEngine(TestMLRunSystem):
                 self.get_remote_pq_source_path(without_prefix=True),
                 self.get_local_pq_source_path(),
             )
+            if not self.test_branch:
+                RemoteSparkRuntime.deploy_default_image()
+            else:
+                sj = new_function(
+                    kind="remote-spark", name="remote-spark-default-image-deploy-temp"
+                )
 
-            sj = new_function(
-                kind="remote-spark", name="remote-spark-default-image-deploy-temp"
-            )
-
-            sj.spec.build.image = RemoteSparkRuntime.default_image
-            sj.with_spark_service(spark_service="dummy-spark")
-            sj.spec.build.commands = [
-                "pip install git+https://github.com/mlrun/mlrun.git@development"
-            ]
-            sj.deploy(with_mlrun=False)
-            get_run_db().delete_function(name=sj.metadata.name)
+                sj.spec.build.image = RemoteSparkRuntime.default_image
+                sj.with_spark_service(spark_service="dummy-spark")
+                sj.spec.build.commands = ["pip install git+" + self.test_branch]
+                sj.deploy(with_mlrun=False)
+                get_run_db().delete_function(name=sj.metadata.name)
+            self.spark_image_deployed = True
 
     def test_basic_remote_spark_ingest(self):
         key = "patient_id"
